rukav in progress

+++ app/Containers/Payment/Data/Migrations/2018_01_01_000001_create_payment_transactions_table.php
@@ -20,7 +20,7 @@ class CreatePaymentTransactionsTable extends Migration
 
             $table->increments('id');
 
-            $table->integer('user_id')->unsigned();
+            $table->unsignedBigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users');
 
             $table->string('gateway');
